# src/Readarr.Api.V1/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Readarr",
    "description": "Readarr API docs - The v3 API docs apply to both v3 and v4 versions of Readarr. Some functionality may only be available in v4 of the Readarr application.",
    "license": {
      "name": "GPL-3.0",
      "url": "https://github.com/Readarr/Readarr/blob/develop/LICENSE"
    },
    "version": "3.0.0"
  },
  "servers": [
    {
      "url": "{protocol}://{hostpath}",
      "variables": {
        "protocol": {
          "default": "http",
          "enum": [
            "http",
            "https"
          ]
        },
        "hostpath": {
          "default": "localhost:8989"
        }
      }
    }
  ],
  "paths": {
...
    "/api/v3/episodefile/editor": {
      "put": {
        "tags": [
          "EpisodeFile"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EpisodeFileListResource"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v3/episodefile/bulk": {
      "delete": {
        "tags": [
          "EpisodeFile"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EpisodeFileListResource"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "EpisodeFile"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EpisodeFileResource"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v3/history/since": {
      "get": {
        "tags": [
          "History"
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "eventType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/EpisodeHistoryEventType"
            }
          },
          {
            "name": "includeSeries",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "includeEpisode",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryResource"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/history/series": {
      "get": {
        "tags": [
          "History"
        ],
        "parameters": [
          {
            "name": "seriesId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "seasonNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "eventType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/EpisodeHistoryEventType"
            }
          },
          {
            "name": "includeSeries",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "includeEpisode",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryResource"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/history/failed/{id}": {
      "post": {
        "tags": [
          "History"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v3/series": {
      "get": {
        "tags": [
          "Series"
        ],
        "parameters": [
          {
            "name": "tvdbId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "includeSeasonImages",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeriesResource"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Series"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeriesResource"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesResource"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlternateTitleResource": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "seasonNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "sceneSeasonNumber": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "sceneOrigin": {
            "type": "string",
            "nullable": true
          },
          "comment": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CustomFormatResource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "includeCustomFormatWhenRenaming": {
            "type": "boolean",
            "nullable": true
          },
          "specifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomFormatSpecificationSchema"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EpisodeFileListResource": {
        "type": "object",
        "properties": {
          "episodeFileIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "languages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Language"
            },
            "nullable": true
          },
          "quality": {
            "$ref": "#/components/schemas/QualityModel"
          },
          "sceneName": {
            "type": "string",
            "nullable": true
          },
          "releaseGroup": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EpisodeFileResource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "seriesId": {
            "type": "integer",
            "format": "int32"
          },
          "seasonNumber": {
            "type": "integer",
            "format": "int32"
          },
          "relativePath": {
            "type": "string",
            "nullable": true
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "dateAdded": {
            "type": "string",
            "format": "date-time"
          },
          "sceneName": {
            "type": "string",
            "nullable": true
          },
          "releaseGroup": {
            "type": "string",
            "nullable": true
          },
          "languages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Language"
            },
            "nullable": true
          },
          "quality": {
            "$ref": "#/components/schemas/QualityModel"
          },
          "customFormats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomFormatResource"
            },
            "nullable": true
          },
          "customFormatScore": {
            "type": "integer",
            "format": "int32"
          },
          "indexerFlags": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "releaseType": {
            "$ref": "#/components/schemas/ReleaseType"
          },
          "mediaInfo": {
            "$ref": "#/components/schemas/MediaInfoResource"
          },
          "qualityCutoffNotMet": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "EpisodeHistoryEventType": {
        "enum": [
          "unknown",
          "grabbed",
          "seriesFolderImported",
          "downloadFolderImported",
          "downloadFailed",
          "episodeFileDeleted",
          "episodeFileRenamed",
          "downloadIgnored"
        ],
        "type": "string"
      },
...
      "HistoryResource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "episodeId": {
            "type": "integer",
            "format": "int32"
          },
          "seriesId": {
            "type": "integer",
            "format": "int32"
          },
          "sourceTitle": {
            "type": "string",
            "nullable": true
          },
          "languages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Language"
            },
            "nullable": true
          },
          "quality": {
            "$ref": "#/components/schemas/QualityModel"
          },
          "customFormats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomFormatResource"
            },
            "nullable": true
          },
          "customFormatScore": {
            "type": "integer",
            "format": "int32"
          },
          "qualityCutoffNotMet": {
            "type": "boolean"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "downloadId": {
            "type": "string",
            "nullable": true
          },
          "eventType": {
            "$ref": "#/components/schemas/EpisodeHistoryEventType"
          },
          "data": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "nullable": true
            },
            "nullable": true
          },
          "episode": {
            "$ref": "#/components/schemas/EpisodeResource"
          },
          "series": {
            "$ref": "#/components/schemas/SeriesResource"
          }
        },
        "additionalProperties": false
      },
...
      "Language": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "MediaInfoResource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "audioBitrate": {
            "type": "integer",
            "format": "int64"
          },
          "audioChannels": {
            "type": "number",
            "format": "double"
          },
          "audioCodec": {
            "type": "string",
            "nullable": true
          },
          "audioLanguages": {
            "type": "string",
            "nullable": true
          },
          "audioStreamCount": {
            "type": "integer",
            "format": "int32"
          },
          "videoBitDepth": {
            "type": "integer",
            "format": "int32"
          },
          "videoBitrate": {
            "type": "integer",
            "format": "int64"
          },
          "videoCodec": {
            "type": "string",
            "nullable": true
          },
          "videoFps": {
            "type": "number",
            "format": "double"
          },
          "videoDynamicRange": {
            "type": "string",
            "nullable": true
          },
          "videoDynamicRangeType": {
            "type": "string",
            "nullable": true
          },
          "resolution": {
            "type": "string",
            "nullable": true
          },
          "runTime": {
            "type": "string",
            "nullable": true
          },
          "scanType": {
            "type": "string",
            "nullable": true
          },
          "subtitles": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "QualityModel": {
        "type": "object",
        "properties": {
          "quality": {
            "$ref": "#/components/schemas/Quality"
          },
          "revision": {
            "$ref": "#/components/schemas/Revision"
          }
        },
        "additionalProperties": false
      },
...
      "ReleaseType": {
        "enum": [
          "unknown",
          "singleEpisode",
          "multiEpisode",
          "seasonPack"
        ],
        "type": "string"
      },
...
      "SeriesResource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "alternateTitles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlternateTitleResource"
            },
            "nullable": true
          },
          "sortTitle": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/SeriesStatusType"
          },
          "ended": {
            "type": "boolean",
            "readOnly": true
          },
          "profileName": {
            "type": "string",
            "nullable": true
          },
          "overview": {
            "type": "string",
            "nullable": true
          },
          "nextAiring": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "previousAiring": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "network": {
            "type": "string",
            "nullable": true
          },
          "airTime": {
            "type": "string",
            "nullable": true
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaCover"
            },
            "nullable": true
          },
          "originalLanguage": {
            "$ref": "#/components/schemas/Language"
          },
          "remotePoster": {
            "type": "string",
            "nullable": true
          },
          "seasons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SeasonResource"
            },
            "nullable": true
          },
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "qualityProfileId": {
            "type": "integer",
            "format": "int32"
          },
          "seasonFolder": {
            "type": "boolean"
          },
          "monitored": {
            "type": "boolean"
          },
          "monitorNewItems": {
            "$ref": "#/components/schemas/NewItemMonitorTypes"
          },
          "useSceneNumbering": {
            "type": "boolean"
          },
          "runtime": {
            "type": "integer",
            "format": "int32"
          },
          "tvdbId": {
            "type": "integer",
            "format": "int32"
          },
          "tvRageId": {
            "type": "integer",
            "format": "int32"
          },
          "tvMazeId": {
            "type": "integer",
            "format": "int32"
          },
          "tmdbId": {
            "type": "integer",
            "format": "int32"
          },
          "firstAired": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastAired": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "seriesType": {
            "$ref": "#/components/schemas/SeriesTypes"
          },
          "cleanTitle": {
            "type": "string",
            "nullable": true
          },
          "imdbId": {
            "type": "string",
            "nullable": true
          },
          "titleSlug": {
            "type": "string",
            "nullable": true
          },
          "rootFolderPath": {
            "type": "string",
            "nullable": true
          },
          "folder": {
            "type": "string",
            "nullable": true
          },
          "certification": {
            "type": "string",
            "nullable": true
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "tags": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "added": {
            "type": "string",
            "format": "date-time"
          },
          "addOptions": {
            "$ref": "#/components/schemas/AddSeriesOptions"
          },
          "ratings": {
            "$ref": "#/components/schemas/Ratings"
          },
          "statistics": {
            "$ref": "#/components/schemas/SeriesStatisticsResource"
          },
          "episodesChanged": {
            "type": "boolean",
            "nullable": true
          },
          "languageProfileId": {
            "type": "integer",
            "format": "int32",
            "readOnly": true,
            "deprecated": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "X-Api-Key": {
        "type": "apiKey",
        "description": "Apikey passed as header",
        "name": "X-Api-Key",
        "in": "header"
      },
      "apikey": {
        "type": "apiKey",
        "description": "Apikey passed as query parameter",
        "name": "apikey",
        "in": "query"
      }
    }
  },
  "security": [
    {
      "X-Api-Key": [ ]
    },
    {
      "apikey": [ ]
    }
  ]
}